backup_20251226_1106/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Document Processing
pypdf==4.0.1
python-multipart==0.0.6

# LangChain Stack (fixo e coerente com 0.1.x)
langchain==0.1.20
langchain-core==0.1.53
langchain-community==0.0.38
langchain-text-splitters==0.0.2

# Integrações LangChain (fixas para evitar resolver-too-deep)
langchain-chroma==0.1.4
langchain-huggingface==0.0.3
langchain-openai==0.0.5
langchain-google-genai==0.0.11

# Vector Store
chromadb==0.4.22

# HF / Embeddings (fixo para evitar explosão transformers/tokenizers)
sentence-transformers==2.6.1
transformers==4.45.2
tokenizers==0.20.1
huggingface-hub==0.23.5

# LLM APIs
google-generativeai==0.4.1
openai==1.10.0

# Search/Retrieval
rank-bm25==0.2.2

# Data Processing
beautifulsoup4==4.12.3
lxml==5.1.0
unidecode==1.3.8

# Database
sqlalchemy==2.0.25

# Configuration
pyyaml==6.0.1

# Data Science
numpy==1.26.3
pandas==2.2.0

# HTTP/Networking
httpx==0.26.0
aiofiles==23.2.1

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0